Update UserService.java
Sensitive Data Exposure - Username exposed while creating New User:

The system already has functionality to hide any parameter value for a function call in case the parameter is sensitive. 
However, it is not being used for the “user” parameter for the UserService.createUser call.

This update will make the the “user” parameter hidden for the UserService.createUser call.